+++ lib/moxl/src/Xec/Payload/Presence.php
@@ -27,14 +27,6 @@ class Presence extends Payload
             $presence = DBPresence::findByStanza($stanza);
             $presence->set($stanza);
 
-            $refreshable = $presence->refreshable;
-            if ($refreshable) {
-                $r = new Get;
-                $r->setAvatarhash($presence->avatarhash)
-                  ->setTo((string)$refreshable)
-                  ->request();
-            }
-
             PresenceBuffer::getInstance()->append($presence, function () use ($presence, $stanza) {
                 if ($presence->muc
                 && isset($stanza->x)) {